Move middelware into middleware folder
