+++ src/controllers/index.ts
@@ -14,7 +14,7 @@ import membershipsRouter from './memberships/_.routing';
 import tokenRouter from './token/_.routing';
 import promoCodesRouter from './promo_codes/promoCodes.router';
 import depositsRouter from './deposits/deposits.router';
-import { checkJwt } from '@/util/jwt';
+import { checkJwt } from '@/middlewares';
 
 const router = express.Router();
