+++ src/middlewares/index.ts
@@ -1,17 +1,9 @@
-import { assertAssetPoolAccess } from './assertAssetPoolAccess';
-import { requireAssetPoolHeader } from './requireAssetPoolHeader';
-import { assertRequestInput } from './assertRequestInput';
-import { errorOutput } from './errorOutput';
-import { errorLogger } from './errorLogger';
-import { errorNormalizer } from './errorNormalizer';
-import { notFoundHandler } from './notFoundHandler';
-
-export {
-    assertAssetPoolAccess,
-    requireAssetPoolHeader,
-    assertRequestInput,
-    errorNormalizer,
-    errorLogger,
-    errorOutput,
-    notFoundHandler,
-};
+export * from './assertAssetPoolAccess';
+export * from './requireAssetPoolHeader';
+export * from './assertRequestInput';
+export * from './errorOutput';
+export * from './errorLogger';
+export * from './errorNormalizer';
+export * from './notFoundHandler';
+export * from './corsHandler';
+export * from './checkJwt';
